# speed up mermin bell benchmark